services/operations_service/requirements.txt
alembic==1.16.5
asyncpg==0.30.0
bcrypt==4.3.0
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.9
pydantic-settings==2.11.0
python-jose==3.5.0
redis==6.4.0
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0
